New Tile Format
Ditched the old idea of having checkboxes and issue definitions
separate. Now, they're together - on the Issue Explained page, there are
now checkboxes for whether or not the issue is of importance to the
patron, a slider for how important it is, and then checkboxes for the
various stances of each issue.